Split ngeo print service
Extract vector handling from the print service.
