+++ src/print/Service.js
@@ -1,12 +1,9 @@
 goog.provide('ngeo.print.Service');
 
-
 goog.require('ngeo');
-goog.require('ngeo.utils');
+goog.require('ngeo.print.VectorEncoder');
 goog.require('ngeo.map.LayerHelper');
 goog.require('ol');
-goog.require('ol.color');
-goog.require('ol.format.GeoJSON');
 goog.require('ol.layer.Image');
 goog.require('ol.layer.Tile');
 goog.require('ol.layer.Vector');
@@ -14,11 +11,7 @@ goog.require('ol.math');
 goog.require('ol.size');
 goog.require('ol.source.ImageWMS');
 goog.require('ol.source.TileWMS');
-goog.require('ol.source.Vector');
 goog.require('ol.source.WMTS');
-goog.require('ol.style.Circle');
-goog.require('ol.style.Icon');
-goog.require('ol.style.RegularShape');
 goog.require('ol.tilegrid.WMTS');
 
 
@@ -83,38 +76,13 @@ ngeo.print.Service = function(url, $http, ngeoLayerHelper) {
    * @private
    */
   this.ngeoLayerHelper_ = ngeoLayerHelper;
+
+  /**
+   * @type {ngeo.print.VectorEncoder}
+   * @private
+   */
+  this.vectorEncoder_ = new ngeo.print.VectorEncoder();
 };
-
-
-/**
- * @enum {string}
- */
-ngeo.print.Service.PrintStyleType = {
-  LINE_STRING: 'LineString',
-  POINT: 'Point',
-  POLYGON: 'Polygon'
-};
-
-
-/**
- * @type {Object.<ol.geom.GeometryType, ngeo.print.Service.PrintStyleType>}
- * @private
- */
-ngeo.print.Service.PrintStyleTypes_ = {
-  'LineString': ngeo.print.Service.PrintStyleType.LINE_STRING,
-  'Point': ngeo.print.Service.PrintStyleType.POINT,
-  'Polygon': ngeo.print.Service.PrintStyleType.POLYGON,
-  'MultiLineString': ngeo.print.Service.PrintStyleType.LINE_STRING,
-  'MultiPoint': ngeo.print.Service.PrintStyleType.POINT,
-  'MultiPolygon': ngeo.print.Service.PrintStyleType.POLYGON
-};
-
-
-/**
- * @const
- * @private
- */
-ngeo.print.Service.FEAT_STYLE_PROP_PREFIX_ = '_ngeo_style_';
 
 
 /**
@@ -216,7 +184,7 @@ ngeo.print.Service.prototype.encodeLayer = function(arr, layer, resolution) {
   } else if (layer instanceof ol.layer.Tile) {
     this.encodeTileLayer_(arr, layer);
   } else if (layer instanceof ol.layer.Vector) {
-    this.encodeVectorLayer_(arr, layer, resolution);
+    this.vectorEncoder_.encodeVectorLayer(arr, layer, resolution);
   }
 };
 
@@ -398,414 +366,6 @@ ngeo.print.Service.prototype.encodeTileWmsLayer_ = function(arr, layer) {
 
 
 /**
- * @param {Array.<MapFishPrintLayer>} arr Array.
- * @param {ol.layer.Vector} layer Layer.
- * @param {number} resolution Resolution.
- * @private
- */
-ngeo.print.Service.prototype.encodeVectorLayer_ = function(arr, layer, resolution) {
-  const source = layer.getSource();
-  goog.asserts.assertInstanceof(source, ol.source.Vector);
-
-  const features = source.getFeatures();
-
-  const geojsonFormat = new ol.format.GeoJSON();
-
-  const /** @type {Array.<GeoJSONFeature>} */ geojsonFeatures = [];
-  const mapfishStyleObject = /** @type {MapFishPrintVectorStyle} */ ({
-    version: 2
-  });
-
-  for (let i = 0, ii = features.length; i < ii; ++i) {
-    const originalFeature = features[i];
-
-    let styleData = null;
-    let styleFunction = originalFeature.getStyleFunction();
-    if (styleFunction !== undefined) {
-      styleData = styleFunction.call(originalFeature, resolution);
-    } else {
-      styleFunction = layer.getStyleFunction();
-      if (styleFunction !== undefined) {
-        styleData = styleFunction.call(layer, originalFeature, resolution);
-      }
-    }
-    const origGeojsonFeature = geojsonFormat.writeFeatureObject(originalFeature);
-    /**
-     * @type {Array<ol.style.Style>}
-     */
-    const styles = (styleData !== null && !Array.isArray(styleData)) ?
-      [styleData] : styleData;
-    goog.asserts.assert(Array.isArray(styles));
-
-    if (styles !== null && styles.length > 0) {
-      let isOriginalFeatureAdded = false;
-      for (let j = 0, jj = styles.length; j < jj; ++j) {
-        const style = styles[j];
-        const styleId = ol.getUid(style).toString();
-        let geometry = style.getGeometry();
-        let geojsonFeature;
-        if (!geometry) {
-          geojsonFeature = origGeojsonFeature;
-          geometry = originalFeature.getGeometry();
-          // no need to encode features with no geometry
-          if (!geometry) {
-            continue;
-          }
-          if (!isOriginalFeatureAdded) {
-            geojsonFeatures.push(geojsonFeature);
-            isOriginalFeatureAdded = true;
-          }
-        } else {
-          let styledFeature = originalFeature.clone();
-          styledFeature.setGeometry(geometry);
-          geojsonFeature = geojsonFormat.writeFeatureObject(styledFeature);
-          geometry = styledFeature.getGeometry();
-          styledFeature = null;
-          geojsonFeatures.push(geojsonFeature);
-        }
-
-        const geometryType = geometry.getType();
-        if (geojsonFeature.properties === null) {
-          geojsonFeature.properties = {};
-        }
-
-        const featureStyleProp = ngeo.print.Service.FEAT_STYLE_PROP_PREFIX_ + j;
-        this.encodeVectorStyle_(
-          mapfishStyleObject, geometryType, style, styleId, featureStyleProp);
-        geojsonFeature.properties[featureStyleProp] = styleId;
-      }
-    }
-  }
-
-  // MapFish Print fails if there are no style rules, even if there are no
-  // features either. To work around this we just ignore the layer if the
-  // array of GeoJSON features is empty.
-  // See https://github.com/mapfish/mapfish-print/issues/279
-
-  if (geojsonFeatures.length > 0) {
-    const geojsonFeatureCollection = /** @type {GeoJSONFeatureCollection} */ ({
-      type: 'FeatureCollection',
-      features: geojsonFeatures
-    });
-    const object = /** @type {MapFishPrintVectorLayer} */ ({
-      geoJson: geojsonFeatureCollection,
-      opacity: layer.getOpacity(),
-      style: mapfishStyleObject,
-      type: 'geojson'
-    });
-    arr.push(object);
-  }
-};
-
-
-/**
- * @param {MapFishPrintVectorStyle} object MapFish style object.
- * @param {ol.geom.GeometryType} geometryType Type of the GeoJSON geometry
- * @param {ol.style.Style} style Style.
- * @param {string} styleId Style id.
- * @param {string} featureStyleProp Feature style property name.
- * @private
- */
-ngeo.print.Service.prototype.encodeVectorStyle_ = function(object, geometryType, style, styleId, featureStyleProp) {
-  if (!(geometryType in ngeo.print.Service.PrintStyleTypes_)) {
-    // unsupported geometry type
-    return;
-  }
-  const styleType = ngeo.print.Service.PrintStyleTypes_[geometryType];
-  const key = `[${featureStyleProp} = '${styleId}']`;
-  if (key in object) {
-    // do nothing if we already have a style object for this CQL rule
-    return;
-  }
-  const styleObject = /** @type {MapFishPrintSymbolizers} */ ({
-    symbolizers: []
-  });
-  object[key] = styleObject;
-  const fillStyle = style.getFill();
-  const imageStyle = style.getImage();
-  const strokeStyle = style.getStroke();
-  const textStyle = style.getText();
-  if (styleType == ngeo.print.Service.PrintStyleType.POLYGON) {
-    if (fillStyle !== null) {
-      this.encodeVectorStylePolygon_(
-        styleObject.symbolizers, fillStyle, strokeStyle);
-    }
-  } else if (styleType == ngeo.print.Service.PrintStyleType.LINE_STRING) {
-    if (strokeStyle !== null) {
-      this.encodeVectorStyleLine_(styleObject.symbolizers, strokeStyle);
-    }
-  } else if (styleType == ngeo.print.Service.PrintStyleType.POINT) {
-    if (imageStyle !== null) {
-      this.encodeVectorStylePoint_(styleObject.symbolizers, imageStyle);
-    }
-  }
-  if (textStyle !== null) {
-    this.encodeTextStyle_(styleObject.symbolizers, textStyle);
-  }
-};
-
-
-/**
- * @param {MapFishPrintSymbolizerPoint|MapFishPrintSymbolizerPolygon} symbolizer MapFish Print symbolizer.
- * @param {!ol.style.Fill} fillStyle Fill style.
- * @private
- */
-ngeo.print.Service.prototype.encodeVectorStyleFill_ = function(symbolizer, fillStyle) {
-  let fillColor = fillStyle.getColor();
-  if (fillColor !== null) {
-    goog.asserts.assert(typeof fillColor === 'string' || Array.isArray(fillColor));
-    fillColor = ol.color.asArray(fillColor);
-    goog.asserts.assert(Array.isArray(fillColor), 'only supporting fill colors');
-    symbolizer.fillColor = ngeo.utils.rgbArrayToHex(fillColor);
-    symbolizer.fillOpacity = fillColor[3];
-  }
-};
-
-
-/**
- * @param {Array.<MapFishPrintSymbolizer>} symbolizers Array of MapFish Print
- *     symbolizers.
- * @param {!ol.style.Stroke} strokeStyle Stroke style.
- * @private
- */
-ngeo.print.Service.prototype.encodeVectorStyleLine_ = function(symbolizers, strokeStyle) {
-  const symbolizer = /** @type {MapFishPrintSymbolizerLine} */ ({
-    type: 'line'
-  });
-  this.encodeVectorStyleStroke_(symbolizer, strokeStyle);
-  symbolizers.push(symbolizer);
-};
-
-
-/**
- * @param {Array.<MapFishPrintSymbolizer>} symbolizers Array of MapFish Print
- *     symbolizers.
- * @param {!ol.style.Image} imageStyle Image style.
- * @private
- */
-ngeo.print.Service.prototype.encodeVectorStylePoint_ = function(symbolizers, imageStyle) {
-  let symbolizer;
-  if (imageStyle instanceof ol.style.Circle) {
-    symbolizer = /** @type {MapFishPrintSymbolizerPoint} */ ({
-      type: 'point'
-    });
-    symbolizer.pointRadius = imageStyle.getRadius();
-    const fillStyle = imageStyle.getFill();
-    if (fillStyle !== null) {
-      this.encodeVectorStyleFill_(symbolizer, fillStyle);
-    }
-    const strokeStyle = imageStyle.getStroke();
-    if (strokeStyle !== null) {
-      this.encodeVectorStyleStroke_(symbolizer, strokeStyle);
-    }
-  } else if (imageStyle instanceof ol.style.Icon) {
-    const src = imageStyle.getSrc();
-    if (src !== undefined) {
-      symbolizer = /** @type {MapFishPrintSymbolizerPoint} */ ({
-        type: 'point',
-        externalGraphic: src
-      });
-      const opacity = imageStyle.getOpacity();
-      if (opacity !== null) {
-        symbolizer.graphicOpacity = opacity;
-      }
-      const size = imageStyle.getSize();
-      if (size !== null) {
-        let scale = imageStyle.getScale();
-        if (isNaN(scale)) {
-          scale = 1;
-        }
-        symbolizer.graphicWidth = size[0] * scale;
-        symbolizer.graphicHeight = size[1] * scale;
-      }
-      let rotation = imageStyle.getRotation();
-      if (isNaN(rotation)) {
-        rotation = 0;
-      }
-      symbolizer.rotation = ol.math.toDegrees(rotation);
-    }
-  } else if (imageStyle instanceof ol.style.RegularShape) {
-    /**
-     * Mapfish Print does not support image defined with ol.style.RegularShape.
-     * As a workaround, I try to map the image on a well-known image name.
-     */
-    const points = /** @type {ol.style.RegularShape} */ (imageStyle).getPoints();
-    if (points !== null) {
-      symbolizer = /** @type {MapFishPrintSymbolizerPoint} */ ({
-        type: 'point'
-      });
-      if (points === 4) {
-        symbolizer.graphicName = 'square';
-      } else if (points === 3) {
-        symbolizer.graphicName = 'triangle';
-      } else if (points === 5) {
-        symbolizer.graphicName = 'star';
-      } else if (points === 8) {
-        symbolizer.graphicName = 'cross';
-      }
-      const sizeShape = imageStyle.getSize();
-      if (sizeShape !== null) {
-        symbolizer.graphicWidth = sizeShape[0];
-        symbolizer.graphicHeight = sizeShape[1];
-      }
-      const rotationShape = imageStyle.getRotation();
-      if (!isNaN(rotationShape) && rotationShape !== 0) {
-        symbolizer.rotation = ol.math.toDegrees(rotationShape);
-      }
-      const opacityShape = imageStyle.getOpacity();
-      if (opacityShape !== null) {
-        symbolizer.graphicOpacity = opacityShape;
-      }
-      const strokeShape = imageStyle.getStroke();
-      if (strokeShape !== null) {
-        this.encodeVectorStyleStroke_(symbolizer, strokeShape);
-      }
-      const fillShape = imageStyle.getFill();
-      if (fillShape !== null) {
-        this.encodeVectorStyleFill_(symbolizer, fillShape);
-      }
-    }
-  }
-  if (symbolizer !== undefined) {
-    symbolizers.push(symbolizer);
-  }
-};
-
-
-/**
- * @param {Array.<MapFishPrintSymbolizer>} symbolizers Array of MapFish Print
- *     symbolizers.
- * @param {!ol.style.Fill} fillStyle Fill style.
- * @param {ol.style.Stroke} strokeStyle Stroke style.
- * @private
- */
-ngeo.print.Service.prototype.encodeVectorStylePolygon_ = function(symbolizers, fillStyle, strokeStyle) {
-  const symbolizer = /** @type {MapFishPrintSymbolizerPolygon} */ ({
-    type: 'polygon'
-  });
-  this.encodeVectorStyleFill_(symbolizer, fillStyle);
-  if (strokeStyle !== null) {
-    this.encodeVectorStyleStroke_(symbolizer, strokeStyle);
-  }
-  symbolizers.push(symbolizer);
-};
-
-
-/**
- * @param {MapFishPrintSymbolizerPoint|MapFishPrintSymbolizerLine|MapFishPrintSymbolizerPolygon}
- *      symbolizer MapFish Print symbolizer.
- * @param {!ol.style.Stroke} strokeStyle Stroke style.
- * @private
- */
-ngeo.print.Service.prototype.encodeVectorStyleStroke_ = function(symbolizer, strokeStyle) {
-  const strokeColor = strokeStyle.getColor();
-  if (strokeColor !== null) {
-    goog.asserts.assert(typeof strokeColor === 'string' || Array.isArray(strokeColor));
-    const strokeColorRgba = ol.color.asArray(strokeColor);
-    goog.asserts.assert(Array.isArray(strokeColorRgba), 'only supporting stroke colors');
-    symbolizer.strokeColor = ngeo.utils.rgbArrayToHex(strokeColorRgba);
-    symbolizer.strokeOpacity = strokeColorRgba[3];
-  }
-  const strokeDashstyle = strokeStyle.getLineDash();
-  if (strokeDashstyle !== null) {
-    symbolizer.strokeDashstyle = strokeDashstyle.join(' ');
-  }
-  const strokeWidth = strokeStyle.getWidth();
-  if (strokeWidth !== undefined) {
-    symbolizer.strokeWidth = strokeWidth;
-  }
-};
-
-
-/**
- * @param {Array.<MapFishPrintSymbolizerText>} symbolizers Array of MapFish Print
- *     symbolizers.
- * @param {!ol.style.Text} textStyle Text style.
- * @private
- */
-ngeo.print.Service.prototype.encodeTextStyle_ = function(symbolizers, textStyle) {
-  const symbolizer = /** @type {MapFishPrintSymbolizerText} */ ({
-    type: 'Text'
-  });
-  const label = textStyle.getText();
-  if (label !== undefined) {
-    symbolizer.label = label;
-    let xAlign = 'c';
-    let yAlign = 'm';
-
-    const olTextAlign = textStyle.getTextAlign();
-    // 'left', 'right', 'center', 'end' or 'start'.
-    if (olTextAlign === 'left' || olTextAlign === 'start') {
-      xAlign = 'l';
-    } else if (olTextAlign === 'right' || olTextAlign === 'end') {
-      xAlign = 'r';
-    }
-
-    const olTextBaseline = textStyle.getTextBaseline();
-    // 'bottom', 'top', 'middle', 'alphabetic', 'hanging' or 'ideographic'
-    if (olTextBaseline === 'bottom') {
-      yAlign = 'l';
-    } else if (olTextBaseline === 'top') {
-      yAlign = 't';
-    }
-    symbolizer.labelAlign = `${xAlign}${yAlign}`;
-
-    const labelRotation = textStyle.getRotation();
-    if (labelRotation !== undefined) {
-      // Mapfish Print expects a string, not a number to rotate text
-      symbolizer.labelRotation = (labelRotation * 180 / Math.PI).toString();
-      // rotate around the vertical/horizontal center
-      symbolizer.labelAlign = 'cm';
-    }
-
-    const fontStyle = textStyle.getFont();
-    if (fontStyle !== undefined) {
-      const font = fontStyle.split(' ');
-      if (font.length >= 3) {
-        symbolizer.fontWeight = font[0];
-        symbolizer.fontSize = font[1];
-        symbolizer.fontFamily = font.splice(2).join(' ');
-      }
-    }
-
-    const strokeStyle = textStyle.getStroke();
-    if (strokeStyle !== null) {
-      const strokeColor = strokeStyle.getColor();
-      goog.asserts.assert(typeof strokeColor === 'string' || Array.isArray(strokeColor));
-      const strokeColorRgba = ol.color.asArray(strokeColor);
-      goog.asserts.assert(Array.isArray(strokeColorRgba), 'only supporting stroke colors');
-      symbolizer.haloColor = ngeo.utils.rgbArrayToHex(strokeColorRgba);
-      symbolizer.haloOpacity = strokeColorRgba[3];
-      const width = strokeStyle.getWidth();
-      if (width !== undefined) {
-        // Width is a stroke, radius a radius, so divide by 2
-        symbolizer.haloRadius = width / 2;
-      }
-    }
-
-    const fillStyle = textStyle.getFill();
-    if (fillStyle !== null) {
-      const fillColor = fillStyle.getColor();
-      goog.asserts.assert(typeof fillColor === 'string' || Array.isArray(fillColor));
-      const fillColorRgba = ol.color.asArray(fillColor);
-      goog.asserts.assert(Array.isArray(fillColorRgba), 'only supporting fill colors');
-      symbolizer.fontColor = ngeo.utils.rgbArrayToHex(fillColorRgba);
-    }
-
-    // Mapfish Print allows offset only if labelAlign is defined.
-    if (symbolizer.labelAlign !== undefined) {
-      symbolizer.labelXOffset = textStyle.getOffsetX();
-      // Mapfish uses the opposite direction of OpenLayers for y axis, so the
-      // minus sign is required for the y offset to be identical.
-      symbolizer.labelYOffset = -textStyle.getOffsetY();
-    }
-
-    symbolizers.push(symbolizer);
-  }
-};
-
-
-/**
  * Return the WMTS URL to use in the print spec.
  * @param {ol.source.WMTS} source The WMTS source.
  * @return {string} URL.
@@ -901,7 +461,9 @@ ngeo.print.Service.createPrintServiceFactory = function($http, ngeoLayerHelper) 
 /**
  * @type {!angular.Module}
  */
-ngeo.print.Service.module = angular.module('ngeoPrint', []);
+ngeo.print.Service.module = angular.module('ngeoPrint', [
+  ngeo.map.LayerHelper.module.name
+]);
 ngeo.print.Service.module.service('ngeoPrintService', ngeo.print.Service);
 ngeo.print.Service.module.factory('ngeoCreatePrint', ngeo.print.Service.createPrintServiceFactory);
 ngeo.module.requires.push(ngeo.print.Service.module.name);
